+++ src/widgets/sidebar/ui/sidebar/Sidebar.tsx
@@ -1,24 +1,9 @@
 import { Link } from 'react-router-dom'
-import { IconName } from '@/shared'
 /* eslint-disable import/no-internal-modules */
 import mainLogo from '@/shared/assets/images/mainLogo.png'
 import { SidebarItem } from '../sidebar-item'
+import { studentItems } from './sidebar.data'
 import cl from './Sidebar.module.scss'
-
-type TSidebarItem = {
-	title: string
-	icon: IconName
-	route: string 
-}
-
-const studentItems: TSidebarItem[] = [
-	{ title: 'Dashboard', icon: 'Dashboard', route: '#' },
-	{ title: 'Courses', icon: 'Courses', route: '#' },
-	{ title: 'Calendar', icon: 'Calendar', route: '#' },
-	{ title: 'Teachers', icon: 'Teachers', route: '#' },
-	{ title: 'Messenger', icon: 'Messenger', route: '#' },
-	{ title: 'Info Portal', icon: 'InfoPortal', route: '#' }
-]
 
 export const Sidebar = () => {
 	return (
