+++ src/widgets/sidebar/ui/sidebar/sidebar.data.ts
@@ -1,0 +1,32 @@
+import { IconName } from '@/shared'
+
+type TSidebarItem = {
+	title: string
+	icon: IconName
+	route: string
+}
+
+export const studentItems: TSidebarItem[] = [
+	{ title: 'Dashboard', icon: 'Dashboard', route: '#' },
+	{ title: 'Courses', icon: 'Courses', route: '#' },
+	{ title: 'Calendar', icon: 'Calendar', route: '#' },
+	{ title: 'Teachers', icon: 'Teachers', route: '#' },
+	{ title: 'Messenger', icon: 'Messenger', route: '#' },
+	{ title: 'Info Portal', icon: 'InfoPortal', route: '#' }
+]
+export const managerItems: TSidebarItem[] = [
+	{ title: 'Dashboard', icon: 'Dashboard', route: '#' },
+	{ title: 'Courses', icon: 'Courses', route: '#' },
+	{ title: 'Calendar', icon: 'Calendar', route: '#' },
+	{ title: 'Teachers', icon: 'Teachers', route: '#' },
+	{ title: 'Messenger', icon: 'Messenger', route: '#' },
+	{ title: 'Info Portal', icon: 'InfoPortal', route: '#' }
+]
+export const mentorItems: TSidebarItem[] = [
+	{ title: 'Dashboard', icon: 'Dashboard', route: '#' },
+	{ title: 'Courses', icon: 'Courses', route: '#' },
+	{ title: 'Calendar', icon: 'Calendar', route: '#' },
+	{ title: 'Teachers', icon: 'Teachers', route: '#' },
+	{ title: 'Messenger', icon: 'Messenger', route: '#' },
+	{ title: 'Info Portal', icon: 'InfoPortal', route: '#' }
+]
